Created a new file to track Entity learning
I had hoped to find documentaiton on entities in the Unity website, but
only found non-code documentaiton. I could easily find documentation
from others, but want to make sure it can be done from the Unity docs
alone. Tomorrow I will read through them more in-depth and determine if
I can create entities only using the documentation that I've found. For
now I bookmarked the page and will get back to it later. This is all an
attempt to get the information straight from Unity and not any third
party.

Changes to be committed:
	new file:   2.md